cbr_ui/requirements.txt
aiohappyeyeballs==2.4.4 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
aiohttp==3.11.11 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
aiosignal==1.3.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
altair==5.5.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
astroid==3.3.8 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
attrs==24.3.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
blinker==1.9.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
cachetools==5.5.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
certifi==2024.12.14 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
charset-normalizer==3.4.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
click==8.1.8 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0" and (sys_platform == "win32" or platform_system == "Windows")
contourpy==1.3.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
cycler==0.12.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
dill==0.3.9 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
flake8==7.1.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
fonttools==4.55.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
frozenlist==1.5.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
gitdb==4.0.11 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
gitpython==3.1.43 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
idna==3.10 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
isort==5.13.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
jinja2==3.1.5 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
joblib==1.4.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
jsonschema-specifications==2024.10.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
jsonschema==4.23.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
kiwisolver==1.4.8 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
markdown-it-py==3.0.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
markupsafe==3.0.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
matplotlib==3.10.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
mccabe==0.7.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
mdurl==0.1.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
multidict==6.1.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
narwhals==1.20.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
nltk==3.9.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
numpy==2.2.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
packaging==24.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pandas==2.2.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pillow==11.0.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
platformdirs==4.3.6 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
plotly==5.24.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
propcache==0.2.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
protobuf==5.29.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pyarrow==18.1.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pycodestyle==2.12.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pydeck==0.9.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pyflakes==3.2.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pygments==2.18.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pylint==3.3.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pymystem3==0.2.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pyparsing==3.2.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
python-dateutil==2.9.0.post0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
python-dotenv==1.0.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
pytz==2024.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
referencing==0.35.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
regex==2024.11.6 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
requests==2.32.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
rich==13.9.4 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
rpds-py==0.22.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
scikit-learn==1.6.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
scipy==1.14.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
seaborn==0.13.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
six==1.17.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
smmap==5.0.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
streamlit==1.41.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
tenacity==9.0.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
threadpoolctl==3.5.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
toml==0.10.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
tomlkit==0.13.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
tornado==6.4.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
tqdm==4.67.1 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
tzdata==2024.2 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
urllib3==2.3.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
watchdog==6.0.0 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0" and platform_system != "Darwin"
wordcloud==1.9.4 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
yarl==1.18.3 ; python_full_version >= "3.11.9" and python_full_version < "4.0.0"
